Create ugly "replaceBlankProps" helper function
This is ugly as fuck, but I don't really care after like 20 hours of attempts to make it look pretty and have proper api.

For now, only issue here is that "replacements" parameter have type Partial<P>, when it must be a union of all P keys of type string.

The reason behind this is to force developer to provide replacements for all string props that can be blank strings (i.e. all string props)