Add dynamic translation for <Nothing> card
Now  default <Nothing> card (the one with id "hidden") is not part of stored in the local storage state.

Instead, stored state contains only cards created by user, and <nothing> card is pushed to the store when lessonsStore (aka classesSore) is initialized.

This allows on-fly translation for <nothing> card, which was impossible in previous approach